optimize module template

+++ tests/test_module_actions.py
@@ -1,54 +1,7 @@
-# import os
-# import shutil
-# from pathlib import Path
-
-# from typer.testing import CliRunner
-
-# from jampy_cli import app
-
-# runner = CliRunner()
-
-
-# def test_create_module():
-#     def assert_ok_and_remove():
-#         assert result.exit_code == 0
-#         assert Path(f"{os.path.abspath(os.getcwd())}/{name}").exists() is True
-#         assert Path(f"{os.path.abspath(os.getcwd())}/{name}/.gitignore").exists() is True
-#         shutil.rmtree(module_dir)
-
-#     def assert_misuse():
-#         assert result.exit_code == 2
-#         assert 'usage' in result.stdout.lower()
-
-#     def assert_exists_if_duplicated():
-#         assert result.exit_code == 0
-#         assert 'already exists' in result.stdout.lower()
-#         shutil.rmtree(module_dir)
-
-#     name = "temp_module"
-#     module_dir = Path(f"{os.path.abspath(os.getcwd())}/{name}")
-#     if module_dir.exists():
-#         shutil.rmtree(module_dir)
-
-#     result = runner.invoke(app, ["module", "create", "-p", name])
-#     assert_ok_and_remove()
-
-#     result = runner.invoke(app, ["module", "c", "--path", name])
-#     assert_ok_and_remove()
-
-#     result = runner.invoke(app, ["module", "c", name])
-#     assert_misuse()
-
-#     result = runner.invoke(app, ["module"])
-#     assert_misuse()
-
-#     result = runner.invoke(app, ["module", "create", "-p", name])
-#     result = runner.invoke(app, ["module", "create", "-p", name])
-#     assert_exists_if_duplicated()
-
 import shutil
 from pathlib import Path
 
+from inflection import underscore
 from typer.testing import CliRunner
 
 from jampy_cli import app
@@ -81,8 +34,8 @@ def test_create_module():
         assert 'already exists' in result.stdout.lower()
         shutil.rmtree(module_dir)
 
-    name = "temp_module"
-    module_dir = Path(f"{get_absolute_cwd()}/{name}")
+    name = "temp-module"
+    module_dir = Path(f"{get_absolute_cwd()}/{underscore(name)}")
     if module_dir.exists():
         shutil.rmtree(module_dir)
 
@@ -108,10 +61,20 @@ def test_create_module():
     result = runner.invoke(app, ["module", "create", "-p", name])
     assert_exit_if_duplicated(module_dir)
 
-    sub_name = 'sub_temp_module'
-    sub_module_dir = module_dir.joinpath(sub_name)
+    name = '../temp-module'
+    module_dir = Path(f"{get_absolute_cwd().parent}/{underscore(Path(name).name)}")
+    if module_dir.exists():
+        shutil.rmtree(module_dir)
     assert_target_not_exists(module_dir)
-    assert_target_not_exists(sub_module_dir)
-    result = runner.invoke(app, ["module", "create", "-p", f"{name}/{sub_name}"])
-    assert_ok(sub_module_dir)
-    shutil.rmtree(module_dir)
+    result = runner.invoke(app, ["module", "create", "-p", name])
+    assert_ok(module_dir)
+
+    pname, name = 'temp-dir', 'temp-module'
+    pdir = Path(f"{get_absolute_cwd()}/{pname}")
+    module_dir = Path(f"{pdir}/{underscore(name)}")
+    if pdir.exists():
+        shutil.rmtree(pdir)
+    assert_target_not_exists(pdir)
+    result = runner.invoke(app, ["module", "create", "-p", f"{pname}/{name}"])
+    assert_ok(module_dir)
+    shutil.rmtree(pdir)
